pyproject.toml
[tool.poetry]
name = "recon-service"
version = "0.1.0"
description = "Reconcilation service"
authors = ["saravanan"]
readme = "README.md"
packages = [{include = "recon_service"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
motor = "^3.0.0"
python-dotenv = "^0.21.0"
pymongo = {extras = ["srv"], version = "^4.2.0"}
python-multipart = "^0.0.5"
pandas = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
